Modified Package Meta, Changed Project Structure
