+++ src/rpdk/guard_rail/core/templates/guard-result-pojo.output
