+++ src/rpdk/guard_rail/rule_library/core/schema-linter-core-arn-rules.guard
